refactor: banner-imageDB-moved-to-seperate-file
