+++ src/components/Banner/Banner.js
@@ -2,6 +2,7 @@ import React from "react";
 import "./Banner.css";
 import { truncateString } from "../../utility";
 import Slider from "react-slick";
+import { bannerImageDb } from "./banner_db";
 
 function Banner() {
   const BANNER_DESCRIPTION = `You’ve probably heard of Lorem Ipsum before – it’s the most-used dummy text excerpt out there. 
@@ -10,49 +11,6 @@ function Banner() {
    design work as it should look when fleshed out with text, because it allows viewers to focus on the design work itself, 
    instead of the text. It’s also a great way to showcase the functionality of programs like word processors, font types,
     and more. `;
-
-  const bannerImageDb = [
-    {
-      id: 1,
-      title: "",
-      description: "",
-      img_1200w: "j-cole-everybody-dies.jpg",
-      img_700w: "j-cole-everybody-dies-700w.jpg",
-      imgPortrait: "j-cole-everybody-dies-portrait.jpg",
-    },
-    {
-      id: 2,
-      title: "",
-      description: "",
-      img_1200w: "lauv-mean-it.jpg",
-      img_700w: "lauv-mean-it-700w.jpg",
-      imgPortrait: "lauv-mean-it-portrait.jpg",
-    },
-    {
-      id: 3,
-      title: "",
-      description: "",
-      img_1200w: "marshmellow-stars.jpg",
-      img_700w: "marshmellow-stars-700w.jpg",
-      imgPortrait: "marshmellow-stars-portrait.jpg",
-    },
-    {
-      id: 4,
-      title: "",
-      description: "",
-      img_1200w: "meditation.jpg",
-      img_700w: "meditation-700w.jpg",
-      imgPortrait: "meditation-portrait.jpg",
-    },
-    {
-      id: 5,
-      title: "",
-      description: "",
-      img_1200w: "lofi.jpg",
-      img_700w: "lofi-700w.jpg",
-      imgPortrait: "lofi-portrait.jpg",
-    },
-  ];
 
   var settings = {
     dots: true,
