+++ src/components/Banner/banner_db.js
@@ -1,0 +1,44 @@
+const bannerImageDb = [
+  {
+    id: 1,
+    title: "",
+    description: "",
+    img_1200w: "j-cole-everybody-dies.jpg",
+    img_700w: "j-cole-everybody-dies-700w.jpg",
+    imgPortrait: "j-cole-everybody-dies-portrait.jpg",
+  },
+  {
+    id: 2,
+    title: "",
+    description: "",
+    img_1200w: "lauv-mean-it.jpg",
+    img_700w: "lauv-mean-it-700w.jpg",
+    imgPortrait: "lauv-mean-it-portrait.jpg",
+  },
+  {
+    id: 3,
+    title: "",
+    description: "",
+    img_1200w: "marshmellow-stars.jpg",
+    img_700w: "marshmellow-stars-700w.jpg",
+    imgPortrait: "marshmellow-stars-portrait.jpg",
+  },
+  {
+    id: 4,
+    title: "",
+    description: "",
+    img_1200w: "meditation.jpg",
+    img_700w: "meditation-700w.jpg",
+    imgPortrait: "meditation-portrait.jpg",
+  },
+  {
+    id: 5,
+    title: "",
+    description: "",
+    img_1200w: "lofi.jpg",
+    img_700w: "lofi-700w.jpg",
+    imgPortrait: "lofi-portrait.jpg",
+  },
+];
+
+export { bannerImageDb };
